+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## next_release
+
+* Added EXIT_ON_ERROR env var - now is possible to control exit_on_errors
+behaviour
+* Added NO_CHECK env var, now is possible not perform restic check operation
+* Added NO_STATS env var, now is possible not collect per backup stats
+* Added backup tag to the metric labels (if tags is present)
+
 ## 1.1.0 (2023/02/02)
 
 * Update Restic 0.15.1